feat(backend): improve k8s cronjob
