+++ hapi/src/routes/index.js
@@ -21,7 +21,6 @@ const checkSignupMethod = require('./signup-method/signup-method.route')
 const getReport = require('./get-report/get-report.route')
 const redeemOffer = require('./redeem-offer/redeem-offer.route')
 const checkEmailVerified = require('./check-email-verified/check-email-verified.route')
-const runMonthly = require('./run-monthly/run-monthly.route')
 const donate = require('./donate/donate.route')
 
 module.exports = [
@@ -48,6 +47,5 @@ module.exports = [
   getReport,
   redeemOffer,
   checkEmailVerified,
-  runMonthly,
   donate
 ]
