--- hapi/src/routes/run-monthly/run-monthly.validation.js
@@ -1,10 +1,0 @@
-const Joi = require('@hapi/joi')
-
-module.exports = {
-  payload: Joi.object({
-    input: Joi.object({
-      account: Joi.string().required(),
-      password: Joi.string().required()
-    })
-  }).options({ stripUnknown: true })
-}
